% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical3.R
\name{and3}
\alias{and3}
\title{Vectorized logical with support for short-circuits}
\usage{
and3(x, y, z = NULL)
}
\arguments{
\item{x, y, z}{Logical vectors. If \code{z} is \code{NULL} it is interpreted as \code{TRUE} or
\code{FALSE} recycled to the length of \code{x}.}
}
\description{
Vectorized logical with support for short-circuits
}
