% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_integer_if_safe.R
\name{as_integer_if_safe}
\alias{as_integer_if_safe}
\title{Coerce from double to integer if safe}
\usage{
as_integer_if_safe(x)
}
\arguments{
\item{x}{A double vector. If not a double vector, it is simply returned without any coercion.}
}
\description{
The same as \code{as.integer(x)} but only if \code{x} consists only of
whole numbers and is within the range of integers.
}
