% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_system_time.R
\name{bench_system_time}
\alias{bench_system_time}
\title{Evaluate time of computation}
\usage{
bench_system_time(expr)
}
\arguments{
\item{...}{Passed to \code{\link[bench]{system_time}}.}
}
\description{
(Used for examples and tests)
}
