% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character2integer.R
\name{character2integer}
\alias{character2integer}
\title{Character to numeric}
\usage{
character2integer(x, na.strings = NULL, allow.double = FALSE, option = 0L)
}
\arguments{
\item{x}{A character vector.}

\item{na.strings}{A set of strings that shall be coerced to \code{NA_integer_}}

\item{allow.double}{\code{logical(1)} If \code{TRUE}, a double vector may be returned.
If \code{FALSE}, an error will be emitted. If \code{NA}, numeric values outside
integer range are coerced to \code{NA_integer_}, silently.}

\item{option}{Control behaviour:
\describe{
\item{0}{Strip commas.}
}}
}
\description{
Character to numeric
}
