% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_logical.R
\name{count_logical}
\alias{count_logical}
\title{Count logicals}
\usage{
count_logical(x)
}
\arguments{
\item{x}{A logical vector.}
}
\value{
A vector of 3 elements: the number of \code{FALSE}, \code{TRUE}, and
\code{NA} values in \code{x}.
}
\description{
Count the number of \code{FALSE}, \code{TRUE}, and \code{NA}s.
}
