% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diam.R
\name{diam}
\alias{diam}
\alias{thinner}
\title{What is the diameter of set of points?}
\usage{
diam(x, nThread = getOption("hutilscpp.nThread", 1L))

thinner(x, width, nThread = getOption("hutilscpp.nThread", 1L))
}
\arguments{
\item{x}{A numeric vector.}

\item{nThread}{Number of threads to use.}

\item{width}{\code{numeric(1)} (For \code{thinner}, the maximum width)}
}
\value{
A single value:
\describe{
\item{\code{diam}}{The difference of \code{minmax(x)}}
\item{\code{thinner}}{Equivalent to \code{diam(x) <= width}}
}
}
\description{
Equivalent to \code{diff(minmax(x))}
}
