% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisible.R
\name{divisible}
\alias{divisible}
\alias{divisible2}
\alias{divisible16}
\title{Divisibility}
\usage{
divisible(x, d, nThread = getOption("hutilscpp.nThread", 1L))

divisible2(x, nThread = getOption("hutilscpp.nThread", 1L))

divisible16(x, nThread = getOption("hutilscpp.nThread", 1L))
}
\arguments{
\item{x}{An integer vector}

\item{d}{\code{integer(1)}. The divisor. Zero is an error.}

\item{nThread}{The number of threads to use.}
}
\value{
Logical vector: \code{TRUE} where \code{x} is divisible by \code{d}.

\code{divisible2},\code{divisible16} are short for (and quicker than)
\code{divisible(x, 2)} and \code{divisble(x, 16)}.
}
\description{
Divisibility
}
