% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_pmaxC}
\alias{do_pmaxC}
\alias{do_pmaxC_dbl}
\alias{do_pmaxC_int}
\alias{do_pmax0}
\alias{do_pmaxIPnum0}
\alias{do_pmaxIPint0}
\title{Internal pmaxC helpers}
\usage{
do_pmaxC_dbl(x, a, in_place = FALSE)

do_pmaxC_int(x, a, in_place = FALSE)

do_pmax0(x, in_place = FALSE)

do_pmaxIPnum0(x)

do_pmaxIPint0(x)
}
\arguments{
\item{x}{A numeric/integer vector.}

\item{a}{A single numeric/integer.}

\item{in_place}{Modify \code{x} in place?}
}
\description{
Internal functions used when the overheads of assertions
would be too expensive. The \code{_IP_} flavours modify in place.
}
