% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_pmaxV}
\alias{do_pmaxV}
\alias{do_pmaxNumNum}
\alias{do_pmaxIntInt}
\title{Parallel maximum in C++}
\usage{
do_pmaxNumNum(x, y)

do_pmaxIntInt(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector, the same length as x.}
}
\value{
The parallel maximum of the input values.
}
\description{
A faster \code{pmax()}.
}
