% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_pminV}
\alias{do_pminV}
\alias{do_pminV_dbl}
\alias{do_pminV_int}
\title{Parallel maximum}
\usage{
do_pminV_dbl(x, y, in_place = FALSE)

do_pminV_int(x, y, in_place = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector, the same length as x.}

\item{in_place}{(bool, default: \code{false}) Modify \code{x} in-place?}
}
\value{
The parallel maximum of the input values.
}
\description{
A faster \code{pmin()}.
}
