% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every_int32.R
\name{every_int}
\alias{every_int}
\title{Every integer}
\usage{
every_int(nThread = getOption("hutilsc.nThread", 1L), na = NA_integer_)
}
\arguments{
\item{nThread}{Number of threads.}

\item{na}{Value for \code{NA_INTEGER}.}
}
\description{
Every integer
}
