% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{extent}
\alias{extent}
\alias{thinner}
\title{How wide is the extent of vector?}
\usage{
extent(x, nThread = getOption("hutilscpp.nThread", 1L))

thinner(x, width, nThread = getOption("hutilscpp.nThread", 1L))
}
\arguments{
\item{x}{A numeric vector.}

\item{nThread}{Number of threads to use.}

\item{width}{\code{numeric(1)} (For \code{thinner}, the maximum width)}
}
\value{
A single value:
\describe{
\item{\code{extent}}{The difference of \code{minmax(x)}}
\item{\code{thinner}}{Equivalent to \code{extent(x) <= width}}
}
}
\description{
Equivalent to \code{diff(minmax(x))}
}
