% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutilscpp-package.R
\docType{package}
\name{hutilscpp-package}
\alias{hutilscpp}
\alias{hutilscpp-package}
\title{hutilscpp: Miscellaneous Functions in C++}
\description{
Provides utility functions that are simply, frequently used, but may require higher performance that what can be obtained from base R. Incidentally provides support for 'reverse geocoding', such as matching a point with its nearest neighbour in another array. Used as a complement to package 'hutils' by sacrificing compilation or installation time for higher running speeds. The name is a portmanteau of the author and 'Rcpp'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hughparsonage/hutilscpp}
  \item Report bugs at \url{https://github.com/hughparsonage/hutilscpp/issues}
}

}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

Authors:
\itemize{
  \item Hugh Parsonage \email{hugh.parsonage@gmail.com}
}

Other contributors:
\itemize{
  \item Simon Urbanek (fastmatch components) [contributor]
}

}
\keyword{internal}
