% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sorted.R
\name{is_sorted}
\alias{is_sorted}
\title{Is a vector sorted?}
\usage{
is_sorted(x, asc = NA)
}
\arguments{
\item{x}{An atomic vector.}

\item{asc}{Single logical. If \code{NA}, the default, a vector is considered
sorted if it is either sorted ascending or sorted descending;
if \code{FALSE}, a vector is sorted only if sorted descending;
if \code{TRUE}, a vector is sorted only if sorted ascending.}
}
\value{
\code{is_sorted} returns \code{TRUE} or \code{FALSE}

\code{isntSorted} returns \code{0} if sorted or the first position
that proves the vector is not sorted
}
\description{
Is a vector sorted?
}
