% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_isna.R
\name{mean_isna}
\alias{mean_isna}
\title{Proportion of missing values}
\usage{
mean_isna(x, do_anyNA = TRUE, nThread = getOption("hutilscpp.nThread", 1L))
}
\arguments{
\item{x}{An atomic vector.}

\item{do_anyNA}{Should \code{anyNA(x)} be executed before an attempt to
count the \code{NA}'s in \code{x} one-by-one? By default, set to \code{TRUE},
since it is generally quicker. It will only be slower when \code{NA} is rare
and occurs late in \code{x}.

Ignored silently if \code{nThread != 1}.}

\item{nThread}{\describe{
\item{\code{nThread}}{Number of threads to use.}
}}
}
\description{
The proportion of missing values in an atomic vector,
equivalent to \code{mean(is.na(x))}.

Introduced in \pkg{hutilscpp} 0.11.0; require at least that version
if your code depends on this function.
}
\examples{
mean_isna(c(1:5, NA))
}
