% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModeC.R
\name{ModeC}
\alias{ModeC}
\title{Most common element}
\usage{
ModeC(
  x,
  nThread = getOption("hutilscpp.nThread", 1L),
  .range_fmatch = 1000000000
)
}
\arguments{
\item{x}{An integer vector.}

\item{nThread}{Number of threads to use.}

\item{.range_fmatch}{If the range of \code{x} differs by more than
this amount, the mode will be calculated via \code{fmatchp}.}
}
\description{
Most common element
}
