% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/pmaxC.R
\name{pmaxC}
\alias{pmaxC}
\alias{pminC}
\title{Parallel maximum}
\usage{
pmaxC(x, a)

pminC(x, a = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{a}{A single numeric value.}

\item{x}{A numeric vector.}

\item{a}{A single numeric value.}
}
\value{
The parallel maximum of the input values. \code{pmax0(x)} is
shorthand for \code{pmaxC(x, 0)}, i.e. convert negative values in \code{x} to 0.

The parallel maximum/minimum of the input values. \code{pmax0(x)} is
shorthand for \code{pmaxC(x, 0)}, i.e. convert negative values in \code{x} to 0.
}
\description{
A faster \code{pmax()}.

Faster \code{pmax()} and \code{pmin()}.
}
\note{
This function will always be faster than \code{pmax(x, a)} when \code{a} is
a single value, but can be slower than \code{pmax.int(x, a)} when \code{x} is short.
Use this function when comparing a numeric vector with a single value.

This function will always be faster than \code{pmax(x, a)} when \code{a} is
a single value, but can be slower than \code{pmax.int(x, a)} when \code{x} is short.
Use this function when comparing a numeric vector with a single value.
}
\examples{
pmaxC(-5:5, 2)

}
