% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmaxV.R
\name{pmaxV}
\alias{pmaxV}
\title{Parallel maximum of equi-length vectors}
\usage{
pmaxV(x, y)
}
\arguments{
\item{x, y}{Numeric vectors of equal length and same type.}
}
\description{
Parallel maximum of equi-length vectors
}
