% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmin.R
\name{pmin0}
\alias{pmin0}
\title{Negative component}
\usage{
pmin0(x, in_place = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{in_place}{(logical, default: \code{FALSE}) Should \code{x}
be modified in-place.}
}
\value{
Same as \code{pmin(x, 0)}.
}
\description{
Negative component
}
\examples{
pmin0(-5:5)

}
