% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poleInaccessibility2.R
\name{poleInaccessibility2}
\alias{poleInaccessibility2}
\title{Find a binary pole of inaccesibility}
\usage{
poleInaccessibility2(x = NULL, y = NULL, DT = NULL, x_range = NULL,
  y_range = NULL)
}
\arguments{
\item{x, y}{Coordinates.}

\item{DT}{A \code{data.table} containing \code{LONGITUDE} and \code{LATITUDE} to define
the \code{x} and \code{y} coordinates.}

\item{x_range, y_range}{Numeric vectors of length-2; the range of \code{x} and \code{y}.
Use this rather than the default when the 'vicinity' of \code{x,y} is different from
the minimum closed rectangle covering the points.}

\item{test_both}{(logical, default: \code{TRUE}) For \code{3}, test both stretching vertically
then horizontally and horizontally then vertically.}
}
\value{
A named vector containing the
\code{xmin}, \code{xmax} and
\code{ymin}, \code{ymax} coordinates of
the largest rectangle of width an integer power of two that is empty.
}
\description{
Find a binary pole of inaccesibility
}
