% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqN_by.R
\name{seqN_by}
\alias{seqN_by}
\alias{seqN_to_N}
\alias{mutate_seqN_N}
\title{Running per group identification}
\usage{
seqN_by(x, check = TRUE)

seqN_to_N(seqN, check = TRUE)

mutate_seqN_N(DT, cols = c("seqN", "N"), check = TRUE)
}
\arguments{
\item{x}{A sorted, integer vector.}

\item{check}{Whether to check whether \code{x} is sorted.}

\item{seqN}{A sequence formed from \code{seqN_by}.}

\item{DT}{A \code{data.table} with at least one \code{key}.}

\item{cols}{A length-two character vector, the names of the \code{seqN}
and \code{N} columns.}
}
\value{
For a sorted, integer \code{x}:

\describe{
\item{\code{seqN_by}}{For every run of constant \code{x} the sequence
\code{1, ..., N} where \code{N} is the size of the group. Equivalent to \code{rowid}
in the \code{data.table} package.}
\item{\code{seqN_to_N}}{Accepts a sequence generated by \code{seqN_by} and
returns the sizes of the original group.}
\item{\code{mutate_seqN_N}}{Accepts a \code{data.table} and appends the
\code{seqN} and \code{N} columns.}
}
}
\description{
Within-group row-id and size of a sorted vector.
}
