% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squish.R
\name{squish}
\alias{squish}
\title{Squish into a range}
\usage{
squish(x, a, b, in_place = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{a, b}{Upper and lower bounds}

\item{in_place}{(logical, default: \code{FALSE}) Should the function operate on \code{x} in place?}
}
\value{
A numeric/integer vector with the values of \code{x} "squished" between \code{a}
and \code{b}; values above \code{b}
replaced with \code{b} and values below \code{a} replaced with \code{a}.
}
\description{
Squish into a range
}
\examples{
squish(-5:5,-1L, 1L)

}
