% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_fmatch.R
\name{unique_fmatch}
\alias{unique_fmatch}
\alias{uniqueN_fmatch}
\title{Distinct elements}
\usage{
unique_fmatch(x, nThread = getOption("hutilscpp.nThread", 1L))

uniqueN_fmatch(x, nThread = getOption("hutilscpp.nThread", 1L))
}
\arguments{
\item{x}{An atomic vector.}

\item{nThread}{Number of threads to use.}
}
\value{
Equivalent to \code{unique(x)} or \code{data.table::uniqueN(x)} respectively.
}
\description{
Using the \code{fastmatch} hash functions, determine
the unique elements of a vector, and the number of distinct elements.
}
