% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which3.R
\name{which3}
\alias{which3}
\title{which of three vectors are the elements (all, any) true?}
\usage{
which3(
  x,
  y,
  z,
  And = TRUE,
  anyNAx = anyNA(x),
  anyNAy = anyNA(y),
  anyNAz = anyNA(z)
)
}
\arguments{
\item{x, y, z}{Logical vectors. Either the same length or length-1.
If any input has length zero, the result is \code{integer(0)}.}

\item{And}{Boolean. If \code{TRUE}, only indices where all of x, y, z
are TRUE are returned; if \code{FALSE}, any index where x, y, z
are TRUE are returned.}

\item{anyNAx, anyNAy, anyNAz}{Whether or not the inputs have \code{NA}.}
}
\description{
which of three vectors are the elements (all, any) true?
}
