% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_first.R
\name{which_first}
\alias{which_first}
\title{Where does a logical expression first occur?}
\usage{
which_first(expr)
}
\arguments{
\item{expr}{An expression, such as \code{x == 2}.}
}
\value{
The same as \code{which.max(expr)} or \code{which(expr)[1]} but returns \code{0L}
when \code{expr} has not \code{TRUE} values.
}
\description{
Where does a logical expression first occur?
}
\details{
The performance benefit of using this function is when the RHS of
\code{expr} is a length-one numeric and the resultant logical
expression is lengthy and expensive.
}
\examples{
x <- rep_len(runif(1e4, 0, 6), 5e8)
bench_system_time(x > 5)
bench_system_time(which(x > 5))
bench_system_time(which.max(x > 5))
bench_system_time(which_first(x > 5))

x <- double(1e8)
bench_system_time(x > 0)
bench_system_time(which(x > 0))
bench_system_time(which.max(x > 0))
bench_system_time(which_first(x > 0))

}
