% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_true_onwards.R
\name{which_true_onwards}
\alias{which_true_onwards}
\title{At which point are all values true onwards}
\usage{
which_true_onwards(x)
}
\arguments{
\item{x}{A logical vector. \code{NA} values are not permitted.}
}
\value{
The position of the first \code{TRUE} value in \code{x} at which all
the following values are \code{TRUE}.
}
\description{
At which point are all values true onwards
}
\examples{
which_true_onwards(c(TRUE, FALSE, TRUE, TRUE, TRUE))

}
