% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichs.R
\name{whichs}
\alias{whichs}
\title{Separated which}
\usage{
whichs(
  exprA,
  .env = parent.frame(),
  nThread = getOption("hutilscpp.nThread", 1L)
)
}
\arguments{
\item{exprA}{An expression. Useful when of the form \code{a <op> b} for
\code{a} an atomic vector. Long expressions are not supported.}

\item{.env}{The environment in which \code{exprA} is to be evaluated.}

\item{nThread}{Number of threads to use.}
}
\value{
Integer vector, the indices of \code{exprA} that return \code{TRUE}.
}
\description{
Same as \code{which(exprA)} where \code{exprA} is a binary
expression.
}
