% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xor2.R
\name{xor2}
\alias{xor2}
\title{Exclusive or}
\usage{
xor2(x, y, anyNAx = TRUE, anyNAy = TRUE)
}
\arguments{
\item{x, y}{Logical vectors.}

\item{anyNAx, anyNAy}{Could \code{x} and \code{y} possibly contain \code{NA} values?
Only set to \code{FALSE} if known to be free of \code{NA}.}
}
\description{
Exclusive or
}
