% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_even.R
\name{are_even}
\alias{are_even}
\title{Are even}
\usage{
are_even(x, check_integerish = TRUE)
}
\arguments{
\item{x}{An integer vector. Double vectors may also be used.}

\item{check_integerish}{(logical, default: \code{TRUE}) Should the
values in \code{x} be checked for non-integer values
if \code{x} is a double vector. If \code{TRUE}
and values are found to be non-integer a warning is emitted.}
}
\value{
For \code{are_even}, a logical vector the same length as \code{x},
\code{TRUE} whenever \code{x} is even.

For \code{which_are_even} the integer positions of even values in \code{x}.
}
\description{
Are even
}
