% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce0.R
\name{coalesce0}
\alias{coalesce0}
\title{Convenience function for coalescing to zero}
\usage{
coalesce0(x, nThread = getOption("hutilscpp.nThread", 1L))
}
\arguments{
\item{x}{An atomic vector.}

\item{nThread}{Number of threads to use.}
}
\value{
Equivalent to \code{hutils::coalesce(x, 0)} for
an appropriate type of zero.
}
\description{
Convenience function for coalescing to zero
}
