% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comma.R
\name{Comma}
\alias{Comma}
\title{Faster version}
\usage{
Comma(x, digits = 0L, big.mark = ",")
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{An integer, similar to \code{round}.}

\item{big.mark}{A single character.}
}
\value{
Similar to \code{prettyNum(round(x, digits), big.mark = ',')}.
}
\description{
Faster version
}
