% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisible.R
\name{divisible}
\alias{divisible}
\alias{divisible16}
\title{Divisible}
\usage{
divisible(x, d, nThread = getOption("hutilscpp.nThread", 1L))

divisible16(x, nThread = getOption("hutilscpp.nThread", 1L))
}
\arguments{
\item{x}{An integer vector}

\item{d}{\code{integer(1)}. The divisor.}

\item{nThread}{The number of threads to use.}
}
\value{
Logical vector: \code{TRUE} where \code{x} is divisible by \code{d}.

\code{divisible16} is short for (and quicker than) \code{divisble(x, 16)}.
}
\description{
Divisible
}
