% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_pmaxC}
\alias{do_pmaxC}
\alias{do_pmax0}
\alias{do_pmaxIPnum0}
\alias{do_pmaxIPint0}
\title{Internal pmaxC helpers}
\usage{
do_pmaxC(x, a)

do_pmax0(x)

do_pmaxIPnum0(x)

do_pmaxIPint0(x)
}
\arguments{
\item{x}{A numeric vector.}

\item{a}{A single numeric.}
}
\description{
Internal functions used when the overheads of assertions
would be too expensive.
}
