% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_pmaxV}
\alias{do_pmaxV}
\alias{do_pmaxNumNum}
\alias{do_pmaxIntInt}
\title{Parallel maximum in C++}
\usage{
do_pmaxNumNum(x, y, in_place = FALSE)

do_pmaxIntInt(x, y, in_place = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector, the same length as x.}

\item{in_place}{(bool, default: \code{false}) Should the function operate on \code{x} in-place?}
}
\value{
The parallel maximum of the input values.
}
\description{
A faster \code{pmax()}.
}
