% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutilscpp-package.R
\docType{package}
\name{hutilscpp-package}
\alias{hutilscpp}
\alias{hutilscpp-package}
\title{hutilscpp: Miscellaneous Functions in C++}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
