% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_constant.R
\name{is_constant}
\alias{is_constant}
\title{Is a vector constant?}
\usage{
is_constant(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
\code{TRUE} or \code{FALSE}. Missing values are considered to
be the same.
}
\description{
Efficiently decide whether an atomic vector is constant; that is,
contains only one value.

Equivalent to

\code{data.table::uniqueN(x) == 1L}

or

\code{forecast::is.constant(x)}
}
