% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModeC.R
\name{ModeC}
\alias{ModeC}
\title{Most common element}
\usage{
ModeC(
  x,
  nThread = getOption("hutilscpp.nThread", 1L),
  .range_fmatch = 1e+09,
  option = 1L
)
}
\arguments{
\item{x}{An atomic vector.}

\item{nThread}{Number of threads to use.}

\item{.range_fmatch}{If the range of \code{x} differs by more than
this amount, the mode will be calculated via \code{fmatchp}.}

\item{option}{\code{integer(1)} Handle exceptional cases:
\describe{
\item{0}{Returns \code{NULL} quietly.}
\item{1}{Returns an error if the mode cannot be calculated.}
\item{2}{Emits a warning if the mode cannot be calculate, falls back to \code{hutils::Mode}}
}}
}
\description{
Most common element
}
\examples{
ModeC(c(1L, 1L, 2L))


}
