% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/pmaxV.R
\name{pmaxV}
\alias{pmaxV}
\title{Parallel maximum in C++}
\usage{
pmaxV(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector, the same length as x.}

\item{x, y}{Numeric vectors of equal length and same type.}
}
\value{
The parallel maximum of the input values.
}
\description{
A faster \code{pmax()}.
}
