% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_range.R
\name{range_rcpp}
\alias{range_rcpp}
\title{Range C++}
\usage{
range_rcpp(x, warn_empty = TRUE)
}
\arguments{
\item{x}{A vector for which the range is desired. Missing values
have no defined behaviour.}

\item{warn_empty}{(logical, default: \code{TRUE})}
}
\value{
A vector, the first two positions give the range and
the next two give the positions where the max and min occur.
}
\description{
Range C++
}
