% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum3.R
\name{sum_and3s}
\alias{sum_and3s}
\alias{sum_or3s}
\title{Sum of logical expressions}
\usage{
sum_and3s(
  exprA,
  exprB,
  exprC,
  ...,
  nThread = getOption("hutilscpp.nThread", 1L),
  .env = parent.frame()
)

sum_or3s(
  exprA,
  exprB,
  exprC,
  ...,
  .env = parent.frame(),
  nThread = getOption("hutilscpp.nThread", 1L)
)
}
\arguments{
\item{exprA, exprB, exprC, ...}{Expressions of the form \code{x <op> y}.
with \code{<op>} one of the standard binary operators.}

\item{nThread}{\describe{
\item{\code{integer(1)}}{Number of threads to use.}
}}

\item{.env}{The environment in which the expressions are to be evaluated.}
}
\value{
Equivalent to \code{sum(exprA & exprB & exprC)} or
\code{sum(exprA | exprB | exprC)} as desired.
}
\description{
Sum of logical expressions
}
