% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which3.R
\name{which3}
\alias{which3}
\title{which of three vectors are the elements (all, any) true?}
\usage{
which3(x, y, z, And = TRUE, prepare = FALSE)
}
\arguments{
\item{x, y, z}{Logical vectors. Either the same length or length-1}

\item{And}{Boolean. If \code{TRUE}, only indices where all of x, y, z
are TRUE are returned; if \code{FALSE}, any index where x, y, z
are TRUE are returned.}

\item{prepare}{(logical, default: \code{FALSE})
Create \code{which(x)}, \code{which(y)}, \code{which(z)}
then determine the intersection.}
}
\description{
which of three vectors are the elements (all, any) true?
}
