% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_true_onwards.R
\name{which_true_onwards}
\alias{which_true_onwards}
\title{At which point are all values true onwards}
\usage{
which_true_onwards(x)
}
\arguments{
\item{x}{A logical vector}
}
\description{
At which point are all values true onwards
}
